Extract swtich map mode button
